[fix]: home allcontents로 수정
